Remove AbstractConfig class
