+++ nullaway/src/main/java/com/uber/nullaway/ErrorProneCLIFlagsConfig.java
@@ -22,20 +22,28 @@
 
 package com.uber.nullaway;
 
+import com.google.auto.value.AutoValue;
+import com.google.common.base.Joiner;
+import com.google.common.base.Preconditions;
 import com.google.common.collect.ImmutableSet;
+import com.google.common.collect.Iterables;
 import com.google.errorprone.ErrorProneFlags;
+import com.google.errorprone.util.ASTHelpers;
+import com.sun.tools.javac.code.Symbol;
 import com.uber.nullaway.fixserialization.FixSerializationConfig;
 import com.uber.nullaway.fixserialization.adapters.SerializationAdapter;
 import java.util.Collections;
 import java.util.LinkedHashSet;
 import java.util.Optional;
 import java.util.Set;
+import java.util.regex.Pattern;
+import javax.annotation.Nullable;
 
 /**
  * provides nullability configuration based on additional flags passed to ErrorProne via
  * "-XepOpt:[Namespace:]FlagName[=Value]". See. http://errorprone.info/docs/flags
  */
-final class ErrorProneCLIFlagsConfig extends AbstractConfig {
+final class ErrorProneCLIFlagsConfig implements Config {
 
   private static final String BASENAME_REGEX = ".*/([^/]+)\\.[ja]ar$";
 
@@ -164,6 +172,75 @@ final class ErrorProneCLIFlagsConfig extends AbstractConfig {
 
   private static final String DEFAULT_URL = "http://t.uber.com/nullaway";
 
+  /**
+   * Packages that we assume have appropriate nullability annotations.
+   *
+   * <p>When we see an invocation to a method of a class outside these packages, we optimistically
+   * assume all parameters are @Nullable and the return value is @NonNull
+   */
+  private final Pattern annotatedPackages;
+
+  /**
+   * Sub-packages without appropriate nullability annotations.
+   *
+   * <p>Used to exclude a particular package that contains unannotated code within a larger,
+   * properly annotated, package.
+   */
+  private final Pattern unannotatedSubPackages;
+
+  /** Source code in these classes will not be analyzed for nullability issues */
+  @Nullable private final ImmutableSet<String> sourceClassesToExclude;
+
+  /**
+   * these classes will be treated as unannotated (don't analyze *and* treat methods as unannotated)
+   */
+  @Nullable private final ImmutableSet<String> unannotatedClasses;
+
+  private final Pattern fieldAnnotPattern;
+  private final boolean isExhaustiveOverride;
+  private final boolean isSuggestSuppressions;
+  private final boolean isAcknowledgeRestrictive;
+  private final boolean checkOptionalEmptiness;
+  private final boolean checkContracts;
+  private final boolean handleTestAssertionLibraries;
+  private final ImmutableSet<String> optionalClassPaths;
+  private final boolean assertsEnabled;
+  private final boolean treatGeneratedAsUnannotated;
+  private final boolean acknowledgeAndroidRecent;
+  private final boolean jspecifyMode;
+  private final ImmutableSet<MethodClassAndName> knownInitializers;
+  private final ImmutableSet<String> excludedClassAnnotations;
+  private final ImmutableSet<String> generatedCodeAnnotations;
+  private final ImmutableSet<String> initializerAnnotations;
+  private final ImmutableSet<String> externalInitAnnotations;
+  private final ImmutableSet<String> contractAnnotations;
+  @Nullable private final String castToNonNullMethod;
+  private final String autofixSuppressionComment;
+  private final ImmutableSet<String> skippedLibraryModels;
+  private final ImmutableSet<String> extraFuturesClasses;
+
+  /** --- JarInfer configs --- */
+  private final boolean jarInferEnabled;
+
+  private final boolean jarInferUseReturnAnnotations;
+  private final String jarInferRegexStripModelJarName;
+  private final String jarInferRegexStripCodeJarName;
+  private final String errorURL;
+
+  /** --- Fully qualified names of custom nonnull/nullable annotation --- */
+  private final ImmutableSet<String> customNonnullAnnotations;
+
+  private final ImmutableSet<String> customNullableAnnotations;
+
+  /**
+   * If active, NullAway will write all reporting errors in output directory. The output directory
+   * along with the activation status of other serialization features are stored in {@link
+   * FixSerializationConfig}.
+   */
+  private final boolean serializationActivationFlag;
+
+  private final FixSerializationConfig fixSerializationConfig;
+
   ErrorProneCLIFlagsConfig(ErrorProneFlags flags) {
     if (!flags.get(FL_ANNOTATED_PACKAGES).isPresent()) {
       throw new IllegalStateException(
@@ -290,5 +367,240 @@ final class ErrorProneCLIFlagsConfig extends AbstractConfig {
       Collections.addAll(combined, flagValue.get().split(DELIMITER));
     }
     return ImmutableSet.copyOf(combined);
+  }
+
+  private static final Pattern getPackagePattern(ImmutableSet<String> packagePrefixes) {
+    // noinspection ConstantConditions
+    String choiceRegexp =
+        Joiner.on("|")
+            .join(Iterables.transform(packagePrefixes, input -> input.replaceAll("\\.", "\\\\.")));
+    return Pattern.compile("^(?:" + choiceRegexp + ")(?:\\..*)?");
+  }
+
+  @Override
+  public boolean serializationIsActive() {
+    return serializationActivationFlag;
+  }
+
+  @Override
+  public FixSerializationConfig getSerializationConfig() {
+    Preconditions.checkArgument(
+        serializationActivationFlag, "Fix Serialization is not active, cannot access it's config.");
+    return fixSerializationConfig;
+  }
+
+  @Override
+  public boolean fromExplicitlyAnnotatedPackage(String className) {
+    return annotatedPackages.matcher(className).matches();
+  }
+
+  @Override
+  public boolean fromExplicitlyUnannotatedPackage(String className) {
+    return unannotatedSubPackages.matcher(className).matches();
+  }
+
+  @Override
+  public boolean treatGeneratedAsUnannotated() {
+    return treatGeneratedAsUnannotated;
+  }
+
+  @Override
+  public boolean isExcludedClass(String className) {
+    if (sourceClassesToExclude == null) {
+      return false;
+    }
+    for (String classPrefix : sourceClassesToExclude) {
+      if (className.startsWith(classPrefix)) {
+        return true;
+      }
+    }
+    return false;
+  }
+
+  @Override
+  public boolean isUnannotatedClass(Symbol.ClassSymbol symbol) {
+    if (unannotatedClasses == null) {
+      return false;
+    }
+    String className = symbol.getQualifiedName().toString();
+    for (String classPrefix : unannotatedClasses) {
+      if (className.startsWith(classPrefix)) {
+        return true;
+      }
+    }
+    return false;
+  }
+
+  @Override
+  public ImmutableSet<String> getExcludedClassAnnotations() {
+    return excludedClassAnnotations;
+  }
+
+  @Override
+  public ImmutableSet<String> getGeneratedCodeAnnotations() {
+    return generatedCodeAnnotations;
+  }
+
+  @Override
+  public boolean isInitializerMethodAnnotation(String annotationName) {
+    return initializerAnnotations.contains(annotationName);
+  }
+
+  @Override
+  public boolean isCustomNullableAnnotation(String annotationName) {
+    return customNullableAnnotations.contains(annotationName);
+  }
+
+  @Override
+  public boolean isCustomNonnullAnnotation(String annotationName) {
+    return customNonnullAnnotations.contains(annotationName);
+  }
+
+  @Override
+  public boolean exhaustiveOverride() {
+    return isExhaustiveOverride;
+  }
+
+  @Override
+  public boolean isKnownInitializerMethod(Symbol.MethodSymbol methodSymbol) {
+    Symbol.ClassSymbol enclosingClass = ASTHelpers.enclosingClass(methodSymbol);
+    if (enclosingClass == null) {
+      return false;
+    }
+    MethodClassAndName classAndName =
+        MethodClassAndName.create(
+            enclosingClass.getQualifiedName().toString(), methodSymbol.getSimpleName().toString());
+    return knownInitializers.contains(classAndName);
+  }
+
+  @Override
+  public boolean isExcludedFieldAnnotation(String annotationName) {
+    return Nullness.isNullableAnnotation(annotationName, this)
+        || (fieldAnnotPattern != null && fieldAnnotPattern.matcher(annotationName).matches());
+  }
+
+  @Override
+  public boolean suggestSuppressions() {
+    return isSuggestSuppressions;
+  }
+
+  @Override
+  public boolean acknowledgeRestrictiveAnnotations() {
+    return isAcknowledgeRestrictive;
+  }
+
+  @Override
+  public boolean checkOptionalEmptiness() {
+    return checkOptionalEmptiness;
+  }
+
+  @Override
+  public boolean checkContracts() {
+    return checkContracts;
+  }
+
+  @Override
+  public boolean handleTestAssertionLibraries() {
+    return handleTestAssertionLibraries;
+  }
+
+  @Override
+  public ImmutableSet<String> getOptionalClassPaths() {
+    return optionalClassPaths;
+  }
+
+  @Override
+  public boolean assertsEnabled() {
+    return assertsEnabled;
+  }
+
+  @Override
+  @Nullable
+  public String getCastToNonNullMethod() {
+    return castToNonNullMethod;
+  }
+
+  @Override
+  public String getAutofixSuppressionComment() {
+    if (autofixSuppressionComment.trim().length() > 0) {
+      return "/* " + autofixSuppressionComment + " */ ";
+    } else {
+      return "";
+    }
+  }
+
+  @Override
+  public boolean isExternalInitClassAnnotation(String annotationName) {
+    return externalInitAnnotations.contains(annotationName);
+  }
+
+  @Override
+  public boolean isContractAnnotation(String annotationName) {
+    return contractAnnotations.contains(annotationName);
+  }
+
+  @Override
+  public boolean isSkippedLibraryModel(String classDotMethod) {
+    return skippedLibraryModels.contains(classDotMethod);
+  }
+
+  @Override
+  public ImmutableSet<String> getExtraFuturesClasses() {
+    return extraFuturesClasses;
+  }
+
+  /** --- JarInfer configs --- */
+  @Override
+  public boolean isJarInferEnabled() {
+    return jarInferEnabled;
+  }
+
+  @Override
+  public boolean isJarInferUseReturnAnnotations() {
+    return jarInferUseReturnAnnotations;
+  }
+
+  @Override
+  public String getJarInferRegexStripModelJarName() {
+    return jarInferRegexStripModelJarName;
+  }
+
+  @Override
+  public String getJarInferRegexStripCodeJarName() {
+    return jarInferRegexStripCodeJarName;
+  }
+
+  @Override
+  public String getErrorURL() {
+    return errorURL;
+  }
+
+  @Override
+  public boolean acknowledgeAndroidRecent() {
+    return acknowledgeAndroidRecent;
+  }
+
+  @Override
+  public boolean isJSpecifyMode() {
+    return jspecifyMode;
+  }
+
+  @AutoValue
+  abstract static class MethodClassAndName {
+
+    static MethodClassAndName create(String enclosingClass, String methodName) {
+      return new AutoValue_ErrorProneCLIFlagsConfig_MethodClassAndName(enclosingClass, methodName);
+    }
+
+    static MethodClassAndName fromClassDotMethod(String classDotMethod) {
+      int lastDot = classDotMethod.lastIndexOf('.');
+      String methodName = classDotMethod.substring(lastDot + 1);
+      String className = classDotMethod.substring(0, lastDot);
+      return MethodClassAndName.create(className, methodName);
+    }
+
+    abstract String enclosingClass();
+
+    abstract String methodName();
   }
 }
